+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Commands `wal/project-test` and `wal/project-install` that work like
+  `project-compile` but store the previously used command in a hash
+  table using the project as the key.
+
 ### Removed
 
 - Package `projectile` in favor of using built-in `project`.